# docs/web/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Admin REST API шины данных (§7 ТЗ). Сессии в Redis +\nAPI-токены (Bearer db_*). SPA по embed.FS отдаётся\nфолбэком на index.html для всех путей не из /api/.",
        "title": "Nexus Web API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/settings/app": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Возвращает merge'нутые значения из БД. Секреты (DSN, password) маскируются.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Текущие dynamic-настройки Sentry/ClickHouse (§14.5).",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/nexus_internal_domain.AppSettings"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Patch-семантика: nil-поля сохраняются как есть, \"***\" в секретах не перезаписывает текущее значение.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Обновить dynamic-настройки Sentry/ClickHouse (§14.5).",
                "parameters": [
                    {
                        "description": "patch",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/nexus_internal_domain.AppSettings"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/settings/clickhouse/test": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Открывает временный conn с merge'нутыми (current + body) настройками, делает Ping, возвращает {ok,latency_ms} или {ok:false,error}. Не сохраняет.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Проверить соединение с ClickHouse по patch'у настроек (§7.10).",
                "parameters": [
                    {
                        "description": "patch",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/nexus_internal_domain.ClickHouseSettings"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/nexus_internal_web_usecase.TestResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/settings/notifications/test": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Шлёт тестовое сообщение в чат с merge'нутыми (current + body) настройками. Маскированный bot_token не используется. Возвращает {ok,latency_ms} или {ok:false,error}. Не сохраняет.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Отправить тестовое уведомление в Telegram с patch'ем настроек (§20.7).",
                "parameters": [
                    {
                        "description": "patch",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/nexus_internal_domain.TelegramSettings"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/nexus_internal_web_usecase.TestResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/settings/public": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Лёгкий read-only эндпоинт для любого авторизованного пользователя: возвращает {public_base_url} для сборки полного адреса узла в UI. Не требует прав admin (в отличие от /api/settings/app).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Публичный адрес приложения (§28, Пункт 1).",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.PublicSettingsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/settings/sentry/test": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Создаёт изолированный sentry.Client, отправляет CaptureMessage, ждёт Flush. Глобальный hub не затрагивается. Возвращает {ok,latency_ms} или {ok:false,error}.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Отправить тестовое событие в Sentry с patch'ем настроек (§7.10).",
                "parameters": [
                    {
                        "description": "patch",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/nexus_internal_domain.SentrySettings"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/nexus_internal_web_usecase.TestResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_web_adapter_in_http.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "password_change_required"
                },
                "error": {
                    "type": "string",
                    "example": "not found"
                }
            }
        },
...
        "internal_web_adapter_in_http.PublicSettingsResponse": {
            "type": "object",
            "properties": {
                "public_base_url": {
                    "type": "string"
                }
            }
        },
...
        "nexus_internal_domain.AppSettings": {
            "type": "object",
            "properties": {
                "clickhouse": {
                    "$ref": "#/definitions/nexus_internal_domain.ClickHouseSettings"
                },
                "general": {
                    "$ref": "#/definitions/nexus_internal_domain.GeneralSettings"
                },
                "notifications": {
                    "$ref": "#/definitions/nexus_internal_domain.NotificationsSettings"
                },
                "sentry": {
                    "$ref": "#/definitions/nexus_internal_domain.SentrySettings"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "description": "user_id, кто последним обновил",
                    "type": "string"
                }
            }
        },
...
        "nexus_internal_domain.ClickHouseSettings": {
            "type": "object",
            "properties": {
                "batch_size": {
                    "type": "integer"
                },
                "buffer_max_size": {
                    "type": "integer"
                },
                "database": {
                    "type": "string"
                },
                "flush_interval_sec": {
                    "type": "integer"
                },
                "host": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "user": {
                    "type": "string"
                },
                "workers": {
                    "type": "integer"
                }
            }
        },
        "nexus_internal_domain.GeneralSettings": {
            "type": "object",
            "properties": {
                "public_base_url": {
                    "description": "PublicBaseURL — публичный адрес, под которым опубликован Web (origin без\nхвостового слеша, напр. https://nexus.example.com). Если задан, UI\nформирует полный адрес узла от него вместо window.location.origin.\nnil/\"\" = не задан (UI берёт origin браузера). Не секрет — Get() не маскирует.",
                    "type": "string"
                }
            }
        },
        "nexus_internal_domain.NotificationsSettings": {
            "type": "object",
            "properties": {
                "telegram": {
                    "$ref": "#/definitions/nexus_internal_domain.TelegramSettings"
                }
            }
        },
        "nexus_internal_domain.SentrySettings": {
            "type": "object",
            "properties": {
                "attach_stacktrace": {
                    "type": "boolean"
                },
                "dsn": {
                    "type": "string"
                },
                "enable_tracing": {
                    "type": "boolean"
                },
                "environment": {
                    "type": "string"
                },
                "level": {
                    "type": "integer"
                },
                "traces_sample_rate": {
                    "type": "number"
                },
                "use": {
                    "type": "boolean"
                }
            }
        },
        "nexus_internal_domain.TelegramSettings": {
            "type": "object",
            "properties": {
                "bot_token": {
                    "type": "string"
                },
                "chat_id": {
                    "type": "string"
                },
                "cron": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                }
            }
        },
...
        "nexus_internal_web_usecase.TestResult": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "latency_ms": {
                    "type": "integer"
                },
                "ok": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiTokenAuth": {
            "description": "\"Bearer db_\u003c32-байт-base64\u003e\". См. §7.14.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CookieAuth": {
            "type": "apiKey",
            "name": "nexus_session",
            "in": "cookie"
        }
    }
}